+++ src/components/Toolbar/media.ts
@@ -1,6 +1,7 @@
 import { default as EmojiButton } from "@tiptapify/extensions/components/media/emoji/Button.vue";
 import { default as LinkButton } from "@tiptapify/extensions/components/media/link/Button.vue";
 import { default as ImageButton } from "@tiptapify/extensions/components/media/image/Button.vue";
+import { default as IframeButton } from "@tiptapify/extensions/components/media/iframe/Button.vue";
 import { default as TableButton } from "@tiptapify/extensions/components/media/table/Button.vue";
 import { default as VideoButton } from "@tiptapify/extensions/components/media/video/Button.vue";
 import { markRaw } from "vue";
@@ -20,6 +21,10 @@ export default {
     {
       name: 'video',
       component: markRaw(VideoButton),
+    },
+    {
+      name: 'iframe',
+      component: markRaw(IframeButton),
     },
     {
       name: 'emoji',